+++ pkg/cmd/init.go
@@ -244,8 +244,8 @@ func writeFile(filePath string, fileContents []byte, forceFileOverwrite bool) er
 func newInitCommand() *cobra.Command {
 	cmd := &cobra.Command{
 		Use:   "init",
-		Short: "Use the init command to initialize the Conjur CLI with a Conjur endpoint.",
-		Long: `Use the init command to initialize the Conjur CLI with a Conjur endpoint.
+		Short: "Initialize the Conjur CLI with a Conjur server",
+		Long: `Initialize the Conjur CLI with a Conjur server.
 
 The init command creates a configuration file (.conjurrc) that contains the details for connecting to Conjur. This file is located under the user's root directory.`,
 		SilenceUsage: true,
@@ -269,7 +269,7 @@ The init command creates a configuration file (.conjurrc) that contains the deta
 	cmd.Flags().BoolP("self-signed", "s", false, "Allow self-signed certificates (insecure)")
 	cmd.Flags().BoolP("insecure", "i", false, "Allow non-HTTPS connections (insecure)")
 	cmd.Flags().Bool("force-netrc", false, "Use a file-based credential storage rather than OS-native keystore (for compatibility with Summon)")
-	cmd.Flags().Bool("force", false, "Force overwrite of existing file")
+	cmd.Flags().Bool("force", false, "Force overwrite of existing configuration file")
 
 	return cmd
 }